trading-arena/requirements.txt
# Trading Arena API Dependencies
# Phase 1 - Foundation

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Cryptography (Ed25519 signatures)
PyNaCl>=1.5.0

# Configuration
python-dotenv>=1.0.0

# HTTP Client (for testing and external calls)
httpx>=0.26.0

# Database (future - commented for Phase 1 skeleton)
# asyncpg>=0.29.0
# sqlalchemy>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
